Move tracking implementation from Blocks-specific to a general cart approach
